+++ src/Helper/InlineScript.php
@@ -16,13 +16,6 @@ namespace Zend\View\Helper;
 class InlineScript extends HeadScript
 {
     /**
-     * Registry key for placeholder
-     *
-     * @var string
-     */
-    protected $regKey = 'Zend_View_Helper_InlineScript';
-
-    /**
      * Return InlineScript object
      *
      * Returns InlineScript helper object; optionally, allows specifying a